Update tests directorie:
