+++ Tests/json_tests/input_flags/ignore.txt
@@ -1,0 +1,3 @@
+events.json
+Always
+Sometimes
