+++ Tools/Ill_handler/Readme.md
@@ -1,0 +1,20 @@
+### Распаковка dll файлов
+
+1. Нам нужен Ildasm и Ilasm
+2. Создаём отдельную папку, в которую помещаем данные программы с их системными файлами, а так же bat файлы из bat_scripts
+3. В файле unpack.bat меняем имя файла, который мы хотим распаковать
+4. Запускаем unpack.bat
+5. Берём получившийся .il файл и копируем в input_il
+6. Запускаем скрипт il_unpacker.py
+7. Результат работы в result_il
+
+
+### Запаковка il в dll
+1. В папку `input_origin_il` помещаем не переведённый распакованный `il.txt` файл
+2. В папку `input_translated_il` помещаем переведённый файл
+3. В папке `input_il` должен быть оригинальный .il файл
+4. Запускаем il_packer.py
+5. Результат в `result_il_bin`
+6. Берём результат и заменяем исходный .il файл в папке, где мы производили операцию распаковки dll файла
+7. Запускаем скрипт Pack.bat, предварительно проверив, что там находится корректное имя файла.
+8. В итоге должен получится переведённый dll файл
