+++ Tools/Ill_handler/bat_scripts/Pack.cmd
@@ -1,0 +1,2 @@
+ilasm Sunless.Game.il /resource:Sunless.Game.res /dll /output:Sunless.Game.dll
+pause
