+++ Tools/Ill_handler/bat_scripts/Unpack.cmd
@@ -1,0 +1,1 @@
+ildasm.exe Sunless.Game.dll /out=Sunless.Game.il
