--- Tools/Json_unpacker/tests/input_flags/flags.txt
@@ -1,60 +1,0 @@
-events.json
-ButtonText
-Description
-Name
-Teaser
-MoveMessage
-
-events-norm.json
-ButtonText
-Description
-Name
-Teaser
-MoveMessage
-
-areas.json
-Description
-MoveMessage
-Name
-
-exchanges.json
-Description
-Name
-
-qualities.json
-ChangeDescriptionText
-Description
-LevelDescriptionText
-Name
-AvailableAt
-
-CombatItems.json
-Description
-Name
-
-Tutorials.json
-Description
-Name
-
-SpawnedEntities.json
-HumanName
-
-Tiles.json
-Description
-HumanName
-Label
-
-Associations.json
-Description
-Name
-Tooltip
-
-CombatAttacks.json
-Description
-
-test.json
-ButtonText
-Description
-Name
-Teaser
-MoveMessage
