--- Tools/Json_unpacker/tests/input_flags/ignore.txt
@@ -1,3 +1,0 @@
-events.json
-Always
-Sometimes
