--- Tools/Json_unpacker/tests/input_json_text/events - копия (2).txt
@@ -1,1 +1,0 @@
->>	Description	Как эта тварь вообще попала на борт корабля?
