--- Tools/Json_unpacker/tests/input_json_text_orig/events - копия (2).txt
@@ -1,1 +1,0 @@
->>	Description	How did that horrible creature get aboard the ship?
