--- Tools/Json_unpacker/tests/result_json_bin/events_not_translated.txt
@@ -1,5 +1,0 @@
-lvl: lvl
-Key: Key
-Value: Value
-Orig: Value
-
