--- Tools/Json_unpacker/tests/test_readfile.txt
@@ -1,1 +1,0 @@
->>	Description	\"И как живется там на дальних берегах?\" [Посещайте и исследуйте острова, чтобы собирать Портовые Отчеты, а затем сдавать их здесь.]
