--- Tools/dll_handler/bat_scripts/Pack.cmd
@@ -1,2 +1,0 @@
-ilasm Sunless.Game.il /resource:Sunless.Game.res /dll /output:Sunless.Game.dll
-pause
