--- Tools/dll_handler/bat_scripts/Unpack.cmd
@@ -1,1 +1,0 @@
-ildasm.exe Sunless.Game.dll /out=Sunless.Game.il
